delete env

+++ controller/topicController.js
@@ -2,14 +2,14 @@ const topics = require('../model/topics');
 const user = require('../model/users');
 const resource = require("../model/resource.model");
 const subscription = require('../model/subscription.model')
+const { v4: uuidv4 } = require('uuid');
 
 
 exports.createTopic = async (req, res) => {
   try {
-
     const { name, visibility, } = req.body;
     const id = req.user.user._id;
-    if (!name || !visibility || !id) {
+    if (!name || !visibility ) {
       return res.status(400).json({ message: "All Fields are Required" });
     }
     const isExist = await topics.findOne({ name });
@@ -20,6 +20,7 @@ exports.createTopic = async (req, res) => {
 
     let topicData = new topics({
       name: name,
+      uuid:uuidv4(),
       visibility: visibility,
       createdby: id,
     });
@@ -89,17 +90,20 @@ exports.getpublictopic = async (req, res) => {
 
 exports.deleteTopic = async (req, res) => {
   const { name } = req.body;
+  if(!name){
+    return res.status(400).json({ message: "Topic name is required" });
+  }
   try {
     const topicdata = await topics.findOne({ name });
-    console.log(topicdata);
     if (!topicdata) {
       return res.status(404).json({ message: 'Topic not found' })
-    } else {
+    } 
+    else {
       await topics.deleteOne({ name: name });
       await resource.deleteMany({ topic: topicdata._id });
       return res.status(200).json({ message: "Topic Delete Successfully And Resource Also deleted" });
     }
-  } catch (error) {
+  }catch (error) {
     return res.status(500).json({ message: "Internal Server Error", error: error.message });
   }
 
@@ -109,6 +113,9 @@ exports.deleteTopic = async (req, res) => {
 exports.editTopic = async (req,res)=>{
   const {name,visibility} = req.body;
   const {id} = req.params;
+  if(!name || !visibility || !id){
+    return res.status(400).json({ message: "Topic name , visibility or  id is required" });
+  }
   try {
    const topicdata = await topics.findOne({ _id: id});
    if(!topicdata){
